Inline FileUploadBottomSheet abstract class
It only had 1 implementation anyway
